feat(pipes): add WmTimeFormatterPipe
This commit adds the WmTimeFormatterPipe to the pipe module. The pipe is responsible for formatting a given number of seconds into hours, minutes, and seconds.

chore(track-download-urls): Add conditional button attribute and observable for UGC selection

This commit adds a conditional button attribute to the download buttons in the track-download-urls.component.html file. It also introduces an observable for checking if UGC (User Generated Content) is selected in the track-download-urls.component.ts file. These changes improve the user experience by enabling or disabling the download buttons based on the UGC selection.