feat(api): add openUgcInHome action oc: 4015
This commit adds the `openUgcInHome` action to the `api.actions.ts` file. This action is dispatched in the `StatusFilterComponent` and `WmHomeComponent` files to update the `ugcHome` property in the API state. The reducer in `api.reducer.ts` handles this action by updating the `ugcHome` property accordingly. Additionally, a new selector named `isUgcHome` is added to retrieve the value of the `ugcHome` property from the API state.