chore: Update user-activity actions, reducer, and selector (#29)
This commit updates the user-activity action file by adding two new actions: startLoader and stopLoader. It also modifies the reducer to handle these actions and update the loading state accordingly. Additionally, the selector file is updated to include a new selector for the loading state.

chore: Remove loading property from auth reducer and selectors

The loading property in the auth reducer and selectors has been removed as it is no longer needed. This change simplifies the code and improves performance.

chore(map): add loader start and stop functionality

This commit adds the functionality to start and stop the loader in the map component. The loader is started when rendering a layer or rendering pois starts, and it is stopped when rendering a layer or rendering pois is done.

Co-authored-by: peppedeka <peppedeka@gmail.com>