feat(ec): add initialization logic to ecTracks effect
This commit adds an initialization check to the ecTracks effect in the EcEffects class. If the action has an "init" property, it will make a request to getQuery and dispatch either ecTracksSuccess or ecTracksFailure based on the response. This ensures that the effect is properly initialized before executing other actions.