feat(ec): handle null values in currentEcPoiProperties selector
This commit updates the `currentEcPoiProperties` selector to handle null values. If `currentEcPoi` is null, the selector will return null instead of throwing an error.

feat(ugc-poi-properties): add UgcPoiPropertiesComponent

This commit adds the `UgcPoiPropertiesComponent` to the project. This component is responsible for displaying and editing properties of a user-generated content (UGC) point of interest (POI). It includes functionality such as viewing and editing form data, displaying media content, and deleting the UGC track associated with the POI.

chore(wm-core.module): import UgcPoiPropertiesComponent

This commit imports and includes the `UgcPoiPropertiesComponent` in the list of module declarations in `wm-core.module.ts`.