chore(conf): add flow line quote text to track properties
This commit adds the functionality to display the flow line quote text in the track properties component. The flow line quote text is calculated based on the altitude of the location and certain threshold values. The text will be displayed dynamically based on the altitude value.

The following files were modified:
- projects/wm-core/src/store/conf/conf.selector.ts
- projects/wm-core/src/store/user-activity/user-activity.selector.ts
- projects/wm-core/src/track-properties/track-properties.component.html
- projects/wm-core/src/track-properties/track-properties.component.ts