chore(conf): add flow line quote text to track properties oc:4740,4745,4739,4741,4727 (#40)
* chore(conf): add flow line quote text to track properties

This commit adds the functionality to display the flow line quote text in the track properties component. The flow line quote text is calculated based on the altitude of the location and certain threshold values. The text will be displayed dynamically based on the altitude value.

The following files were modified:
- projects/wm-core/src/store/conf/conf.selector.ts
- projects/wm-core/src/store/user-activity/user-activity.selector.ts
- projects/wm-core/src/track-properties/track-properties.component.html
- projects/wm-core/src/track-properties/track-properties.component.ts

* chore: Remove unused method and update URL handling in TrackPropertiesComponent

The `getFlowPopoverText` method was removed as it was no longer being used. Additionally, the URL handling in the `onLocationHover` method was updated to use the `_urlHandlerSvc.updateURL` function. This ensures that the track property is set to undefined when hovering over a location.