chore: Update ugc-track-data.component.html
- Added *ngIf conditions to display track data only if the values are greater than 0.

chore: Refactor getSlope method to calculate slope from geometry

This commit refactors the getSlope method in the GeoutilsService class to calculate the slope using the geometry of the track. It introduces a new private method, _getSlopeFromGeometry, which calculates the slope based on the altitude difference between consecutive coordinates in the LineString geometry. This change improves code readability and maintainability.