chore: Update track-useful-urls.component.ts
Refactor the code to ensure that the first value is always converted to a string before removing any spaces.

chore(pipe): Add support for translating numbers in WmTransPipe

This commit adds a condition to check if the value passed to the pipe is a number. If it is, the number is converted to a string before being translated using the parent TranslatePipe class. This allows for translating numbers in addition to strings in the WmTransPipe.