feat: Add DownloadsEcTrackBox component oc:4827
Added a new component called DownloadsEcTrackBox to the box module. This component is responsible for displaying a box with downloads related information. It includes an HTML template and SCSS styling file.

The DownloadsEcTrackBox component has been added to the list of box components in the module declaration.

A new HTML template file has been created for the DownloadsEcTrackBox component, which includes a div element with a class "wm-box" and some child elements for displaying content.

A new SCSS styling file has been created for the DownloadsEcTrackBox component, which includes styles for positioning, sizing, and coloring of the box.

The DownloadsEcTrackBox component also includes a TypeScript file that defines its behavior. It extends the BaseBoxComponent class and implements an openDownloads method that dispatches an action to open downloads. The defaultPhotoPath property is set to 'assets/images/profile/downloads.webp'.

In addition to these changes, there are also some minor updates in other files related to conditional rendering based on certain conditions like offline status or download opened status.