chore: Sanitize object values in CameraService
This commit adds a private method `_sanitizeObjectValues` to the `CameraService` class. This method is responsible for sanitizing object values by replacing control characters with '?' and converting all other types to strings. The method is used in the `getPhotos` method to sanitize the `exif` property of each photo before returning them.