chore: Update user-activity.selector.ts and track-properties.component.html oc: 5015 (#66)
* chore: Update user-activity.selector.ts and track-properties.component.html

- Refactored the flowLineQuoteText selector in user-activity.selector.ts to return an object with the flow_line_quote_orange, flow_line_quote_red, and html properties.
- Updated track-properties.component.html to use the new flowLineQuoteHtml$ observable instead of flowLineQuoteText$.
- Added a new observable, flowLineQuoteHtml$, in track-properties.component.ts that translates the html property from the flowLineQuoteText object and replaces placeholders with actual values.

chore: Update localization files

- Updated translations for German, English, Spanish, French, and Italian languages.
- Added new translations for various phrases and terms.
- Improved accuracy and clarity of existing translations.

* chore(user-activity): Add flow line quote text to user activity selector

This commit adds the flow line quote text to the user activity selector in the `user-activity.selector.ts` file. The flow line quote text is generated based on the altitude and flow line values, and it is returned as an object of type `WmSlopeChartFlowLineQuote`. The HTML for the flow line quote text is also updated to include dynamic values for orange and red altitudes.

fix(track-properties): Update import statement in track properties component

This commit updates the import statement for `LangService` in the `track-properties.component.ts` file. The import statement was incorrectly formatted, causing a linting error.