chore: Remove unused code in AuthEffects oc:5027
The `_createErrorAlert` method was returning `null`, which is unnecessary. This commit removes the unused code and returns a promise to create an alert using `_alertCtrl.create()`.

chore: Update error handling in auth.effects.ts

Refactor the code to improve error handling in the AuthEffects class. Instead of directly passing the error message to `_createErrorAlert`, assign it to a variable `errorMessage` for better readability and fallback to a default value if the error message is not available.

chore(i18n): update translations for error messages

- Updated translations for error messages in multiple languages.
- Added translations for "Un utente è già stato registrato con questa email.", "L'email inserita non è corretta. Per favore, riprova.", "La password inserita non è corretta. Per favore, riprova.", and "Errore" in the following languages: German, English, Spanish, French, Italian, Portuguese, and Albanian.

BREAKING CHANGE: None

fix(auth): handle error message in AuthEffects

- Updated the code to handle error messages in the AuthEffects file.
- Fixed an issue where the error object was not properly checked for 'Unauthorized' error.
- Refactored the code to use optional chaining operator for better readability.