chore: Add WmLayerFeaturesCounterBadgeComponent
This commit adds the `WmLayerFeaturesCounterBadgeComponent` to the box module. It also includes changes to the layer-box component and home-landing component templates to use this new component. Additionally, a new scss file and html template are added for the `WmLayerFeaturesCounterBadgeComponent`. The `layerFeaturesCount` selector is updated in the ec.selector file to calculate the counts of features for each layer.