feat: Add action to set map details status
This commit adds a new action `setMapDetailsStatus` to the user-activity.action.ts file. The action allows setting the map details status with options such as 'open', 'onlyTitle', 'none', 'background', or 'toggle'.

chore: Update user-activity.reducer.ts

In this commit, the user-activity.reducer.ts file is updated to include a new type `mapDetailsStatus` which represents the different status options for map details. The initial state of `mapDetailsStatus` is set to 'none'.

chore: Update user-activity.selector.ts

The user-activity.selector.ts file is modified in this commit to include a new selector `mapDetailsStatus`. This selector retrieves the current value of `mapDetailsStatus` from the user activity state.

These changes enable better management and control of map details in the user activity feature.