chore(user-activity): add function to extract filter taxonomies
This commit adds a new function called `extractFilterTaxonomies` to the user-activity.reducer.ts file. This function is used to extract filter taxonomies from a given layer object. It filters out any taxonomies with null identifiers and maps them into an array of strings. The extracted filter taxonomies are then used in the setLayer reducer to update the poisSelectedFilterIdentifiers state based on the selected layer's taxonomies. Additionally, when a layer is set to null, the previous layer's filter taxonomies are removed from the poisSelectedFilterIdentifiers state.

Co-authored-by: Z