chore(user-activity): add backOfMapDetails action
This commit adds the `backOfMapDetails` action to the `user-activity.action.ts` file. This action is triggered when the user wants to go back from the map details view. It updates the URL and sets the map details status to 'background'.

feat(user-activity): handle backOfMapDetails action in effects

This commit adds an effect in the `user-activity.effects.ts` file to handle the `backOfMapDetails` action. It checks the current query parameters and updates the URL accordingly. If there are related POIs or UGC POIs, it removes them from the URL. Otherwise, it removes track and UGC track parameters from the URL and sets the map details status to 'background'.

feat(user-activity): update mapDetailsStatus in reducer

This commit updates the `mapDetailsStatus` state in the `user-activity.reducer.ts` file based on the `setMapDetailsStatus` action. The new status is set according to what is passed in as a parameter.

chore(user-activity): remove unused import

This commit removes an unused import statement from the `user-activity.reducer.ts` file.

chore(user-activity): update initial state of mapDetailsStatus

This commit updates the initial state of `mapDetailsStatus` in the `user-activit