chore: Add features-in-viewport component
This commit adds the features-in-viewport component, which displays a list of features in a slide show format. The component template and styles are also included.

chore: Update geobox-map.component.html and geobox-map.component.ts

- Added class "show-bottom-features" to the div element in geobox-map.component.html
- Added input property "enableFeaturesInViewport$" and event listener "featuresInViewport()" in geobox-map.component.ts

chore: Update user-activity actions and reducer

This commit updates the user-activity.actions.ts file by adding a new action called "wmMapFeaturesInViewport". It also updates the user-activity.reducer.ts file by adding a new case for this action, which sets the "featuresInViewport" property in the state. Additionally, it adds new selectors in the user-activity.selector.ts file to retrieve the features in viewport and check if there are any features in viewport. Finally, it adds a selector to enable features in viewport based on the map details status.

chore: Update user-activity.action.ts, user-activity.effects.ts, and user-activity.reducer.ts

- Add import for IHIT from '@wm-core/types/elastic'
- Modify wmMapFeaturesInViewport action to accept featureIds instead of features
- Add wmMapFeaturesInViewportSuccess and wmMapFeaturesInViewportFailure actions
- Implement featuresInViewport$ effect in user-activity.effects.ts to fetch features based on featureIds and handle success and failure cases
- Update UserActivityState interface in user-activity.reducer.ts to use IHIT type for featuresInViewport property
- Update the reducer logic for wmMapFeaturesInViewport action to set the featuresInViewport state with the response from the API call

chore: Update features-in-viewport component

- Replace TODO comment with wm-search-box component
- Add data binding and click event to wm-search-box
- Remove unnecessary CSS styles for .wm-features-in-viewport-feature
- Update constructor to include UrlHandlerService dependency
- Add setTrack method to handle click event in wm-search-box

chore: Update geobox-map.component.ts

- Import FeatureLike from 'ol/Feature'
- Modify featuresInViewport method to extract feature ids and dispatch wmMapFeaturesInViewport action with the extracted ids

chore(ec): add support for filtering by trackIds

This commit adds the ability to filter tracks in the EcService by their trackIds. The trackIds parameter is now accepted and used to construct the query string for filtering.