chore(embedded-element): add embedded element component (#91)
This commit adds a new component called WmEmbeddedElementComponent. This component is responsible for rendering an embedded element with a title and HTML content. It uses Angular's DomSanitizer to bypass security and display the HTML content safely. The component also includes styling for the title and the embedded element container.

The WmEmbeddedElementComponent is imported and declared in the WmCoreModule, making it available for use in other parts of the application.

The commit also includes an example usage of the WmEmbeddedElementComponent in the ec.selector.ts file, where an embedded element with a title and HTML content is added to a feature object.

This change enhances the functionality of the application by allowing the display of embedded elements with dynamic content.

chore: Remove WmEmbeddedElementComponent and update InnerHtmlComponent with title input field. Add styling for title in InnerHtmlComponent.