chore(slider-filter): ✨ emit lastFilterTypeEvt for both number and range values oc:5634 (#106)
* chore(slider-filter): ✨ emit lastFilterTypeEvt for both number and range values

This update ensures that the `lastFilterTypeEvt` event is emitted when the slider filter's value changes, regardless of whether the value is a single number or a range. This provides more consistent behavior in tracking filter changes.

feat(user-activity): ✨ add effect to set home result tab when UGC is opened

Added a new effect `setHomeResultTabToTrackWhenOpenUgc$` that updates the home result tab to 'tracks' when the `openUgc` action is triggered. This ensures that the user interface reflects the appropriate context when user-generated content is accessed.

Additionally, the effect `setHomeResultTabWhenLastFilterTypeChanged$` was updated to handle any non-null `lastFilterType`, providing dynamic tab selection based on the last filter type used. Also modified `setHomeResultTabWhenTracksCountIsZero$` to skip the initial emission.

* refactor(slider-filter): ♻️ remove duplicate event emission

Moved the `lastFilterTypeEvt.emit('tracks')` call outside the conditional block to remove redundancy.

chore(user-activity): 📝 add comment for `skip` usage

Added a comment explaining the purpose of `skip(1)` in the `setHomeResultTabWhenTracksCountIsZero$` effect to improve code clarity.