chore(geolocation): 🌐 integrate current location with store oc:5741 (#113)
* chore(geolocation): 🌐 integrate current location with store

Added functionality to dispatch the current location to the store using NgRx. This allows for better state management of the geolocation data.

- Removed unused `defaultIfEmpty` import from `rxjs/operators`.
- Added `setCurrentLocation` action in `user-activity.action.ts`.
- Updated `UserActivityState` and `userActivityReducer` to include `currentLocation`.
- Modified selectors to use the current location from the store.
- Updated `GeolocationService` to dispatch the current location to the store when it changes.

TODO: Refactor to use the `currentLocation` from the store instead of `onLocationChange`.

* chore(get-directions): ✨ implement modal for direction options

Refactored the GetDirectionsComponent to use a modal for selecting either the start point or nearest point for directions. Removed old logic for opening Google Maps directly with coordinates and introduced a new action flow with startGetDirections and getDirections actions. Added localization entries for new text strings used in the modal. Created a new ModalGetDirectionsComponent to encapsulate the modal UI and logic. Updated UserActivityEffects to handle the new modal interaction and dispatch the appropriate actions based on user selection.