chore(form): ✨ add conditional display of form field oc:5247 (#118)
* chore(form): ✨ add conditional display of form field

Updated the form component to conditionally display the form field based on the number of available forms. The form field will only be shown if there is more than one form available. This change improves the user interface by preventing unnecessary form field displays when there is only a single form option.

* chore(form): ✨ add fallback UI for single form type selection

Added a fallback UI for cases where there is only one form type available. The form will now display a title with the form type if only one is present, instead of showing a select dropdown. Updated the HTML template to include an `ng-template` for this scenario and added corresponding styling in the SCSS file.

* style(form): 💄 adjust ion-title font size and padding

Updated the `ion-title` within the `wm-form` component to have a font size of 18px and adjusted the padding from 10px to 8px. This change enhances the styling consistency across the form component.

* feat(localization): 🌐 add translation for 'Tipo di form' in multiple languages

Added the translation for 'Tipo di form' in German, English, Spanish, French, Italian, Portuguese, and Albanian localization files. This enhances the application's multilingual support by providing a consistent translation for 'Form type' across various languages.