refactor(meta): 🔄 replace confGeohubId with EnvironmentService for theme application (#121)
Removed the usage of `confGeohubId` observable and replaced it with `shardName` and `appId` from `EnvironmentService`. This change simplifies the theme application logic by directly using the environment service to retrieve the necessary identifiers for constructing the theme stylesheet URL.

Additionally, added a `shardName` getter in the `EnvironmentService` to expose the private `_shardName` property for external usage.