chore(register): ✨ add unified privacy policy link handling oc:5981 (#132)
* chore(register): ✨ add unified privacy policy link handling

Introduce a new constant `DEFAULT_PRIVACY_POLICY_URL` for the default privacy policy URL. Updated the register component to handle privacy policy links more efficiently by consolidating the logic into a single method `openPrivacyPolicy`. This method now checks if a custom privacy HTML is available and opens it in a modal, or defaults to opening the privacy URL in a new tab. Updated localization files to reflect changes in privacy and terms acceptance message.

- Added a new file for constants: `links.ts`, which includes the default privacy policy URL.
- Updated localization files across different languages to include the new message about accepting privacy and terms.
- Simplified the HTML structure in `register.component.html` for displaying the privacy policy acceptance message.
- Adjusted styling in `register.component.scss` for better presentation of the policy link.

* refactor(register): ♻️ simplify modal presentation logic

Updated the subscription logic to directly present the modal if it exists, eliminating the need for a switchMap operator. This refactor streamlines the code and enhances readability.

* style(register): 💄 simplify subscribe callback in register component

Refactor the subscribe callback to a single line by removing unnecessary braces and line breaks. This improves code readability and maintains the same functionality.