fix(geobox-map): 🐛 correct typo in variable names and add new functionality oc:6003 (#134)
* fix(geobox-map): 🐛 correct typo in variable names and add new functionality

Corrected the typo in variable names from "Recoder" to "Recorder" across the component HTML and TypeScript files. Introduced new observables and actions for enabling track and POI recorder panels. Updated selectors and reducers accordingly to manage the new state variables and ensure proper functionality. This refactor improves code readability and functionality by accurately representing the feature's purpose and enhancing the component's capabilities.

* refactor(geobox-map): ♻️ optimize track recording logic with takeUntil

Refactored the track recording logic to utilize `switchMap` and `takeUntil` for better management of the subscription lifecycle. The `enableTrackRecorderPanel$` observable now uses `takeUntil` to automatically unsubscribe when the track recorder is disabled, preventing potential memory leaks and improving code readability.

This change enhances the efficiency of the component by ensuring that the track recording logic only runs when necessary and cleans up resources when the recorder is not enabled.

* refactor(geobox-map): ♻️ remove redundant distinctUntilChanged operator

The `distinctUntilChanged` operator was removed from the `currentPosition$` observable as its functionality was not necessary in this context. The equality check for latitude and longitude was deemed redundant, potentially simplifying the observable stream and improving performance.